vBooking/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-bootstrap-datepicker-plus==3.0.5
image==1.5.33
Pillow==8.3.2
pytz==2021.1
qrcode==7.3
six==1.16.0
sqlparse==0.4.2
